Added support for exp, nbf, iss, and aud claims.
These are registered claims for the JWT spec, so just adding support
for this part of the spec as part of the signature verification
process.

If exp or nbf are in the payload, verification will fail if the token
is expired or not yet active, unless ignoreExpiration is true, in which
case expiration will be ignored. These payloads are expected to be
standard Unix epoch times, according to the spec.

The issuer (iss) and audience (aud) values will be ignored in the
payload unless you provide an issuer and/or audience to the CFC in the
constructor/init.